Рабочая версия вольтдобавочной УКУшки для БПСов с новой аналоговой частью от Стараковсого. Работает с БПСом под ПО FAT42_dev ветка VD_FAST_STAR . Тэг рабочего коммита и там и здесь ЭлектроКонцепт_Release.

+++ SOFT/UKU/mess.c
@@ -24,7 +24,7 @@ for(i=0;i<MESS_DEEP;i++)
 }
 
 //-----------------------------------------------
-void mess_send(char _mess, short par0, short par1, char _time)
+void mess_send(char _mess, int par0, int par1, char _time)
 {
 char i;
 i=0;
